requirements.txt
black==24.10.0
django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==2.2.1
pandas==2.2.3
pymongo==4.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
ruff==0.8.4
scikit-learn==1.6.0
